Backward-compatible Rails 4 route mapper compatibility
There has been discussion in #42 of moving to a rack-based routing
solution, but that has not been resolved, and Rails 4 is out now.
This change would also require a major version bump, thus leaving
1.x incompatible with Rails 4.

This patch adds the required :via option for Rails >= version 4,
but not for Rails 3, which does not support (nor need) :via.